bookpl/arcbook_recognize/requirements.txt
easydict==1.13
mpmath==1.3.0
mxnet-cu116==1.9.1
networkx==3.2.1
numpy==1.23.1
opencv-python==4.10.0.82
pandas==2.2.2
scikit-learn==1.5.0
scipy==1.8.1
tensorboard==2.17.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
wandb==0.17.4
boto3
sagemaker
fastapi
pydantic
uvicorn
async_generator
faiss-cpu
absl-py==1.4.0
asttokens==2.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
exceptiongroup==1.1.2
executing==1.2.0
filelock==3.12.2
fonttools==4.41.0
fsspec==2023.9.1
grpcio==1.57.0
idna==3.4
imutils==0.5.4
ipython==8.18.1
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pillow==9.5.0
prompt_toolkit==3.0.41
protobuf==4.23.4  # Updated for compatibility
psutil==5.9.5
ptyprocess==0.7.0
pure_eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytictoc==1.5.3
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tensorboard-data-server==0.7.0
thop==0.1.1.post2209072238
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.7
onnxruntime-gpu
onnx
ultralytics
pytesseract
anylabeling
pyqt5
pyarrow==14.0.2
umap-learn
